+++ src/Lorentz/Contracts/TZBTC.hs
@@ -15,11 +15,11 @@ module Lorentz.Contracts.TZBTC
   , TZBTCParameter
   , TZBTCStorage
   , UpgradeParameters
+  , OriginationParameters (..)
   , tzbtcContract
   , fromFlatParameter
   , mkEmptyStorageV0
   , migrationScripts
-  , originationParams
   , toSafeParam
   , tzbtcContractRouter
   , tzbtcDoc
@@ -31,7 +31,6 @@ import Lorentz.Contracts.TZBTC.FlatParameter
 import Lorentz.Contracts.TZBTC.Preprocess
 import Lorentz.Contracts.TZBTC.Types as Types
 import Lorentz.Contracts.TZBTC.V0
-import Lorentz.Contracts.TZBTC.V1
 
 -- Implementation
 ----------------------------------------------------------------------------